Completed basic I2C supported - not tested yet (need to test ADC10 and Wire soon on G2553).

+++ Implementations/msp430_value/platform.h
@@ -39,8 +39,8 @@ enum PortselMode {
 
 typedef const volatile uint8_t & u8_CSFR;
 typedef volatile uint8_t & u8_SFR;
-typedef const volatile uint16_t & u16_CSFR;
-typedef volatile uint16_t & u16_SFR;
+typedef const volatile unsigned int & u16_CSFR;
+typedef volatile unsigned int & u16_SFR;
 
 
 // C linkage (no function overloading)
